잔 버그 수정

+++ Assets/Script/CreatureProperty.cs
@@ -48,7 +48,7 @@ public class CreatureProperty {
 
 	public float getExpRemainRatio()
 	{
-		return Exp/MaxExp;
+		return (float)Exp/MaxExp;
 	}
 
 	public int MaxExp
